+++ src/main/java/digidoggy/scrapdun/Test.java
@@ -1,5 +1,7 @@
 package digidoggy.scrapdun;
 
+import digidoggy.scrapdun.combat.CombatMechanics;
+
 import java.util.Scanner;
 
 
@@ -9,6 +11,8 @@ public class Test {
 
     public static void main(String[] args) {
 
+        CombatMechanics.health(50);
+//        System.out.println("----- -\\/-\\/- /\\/\\/\\/");
     }